+++ apps/app.py
@@ -29,35 +29,5 @@ engine = SQLAlchemy(app)
 def main():
     return render_template('kintai-input.html')
 
-@app.route('/kintai/input')
-def input():
-    return render_template('kintai-input.html')
-
-@app.route('/kintai/search')
-def search():
-    return render_template('kintai-search.html', list=[
-        {
-            "date": "01",
-            "start": "09:00",
-            "end": "18:00",
-            "break": "1:00",
-            "remark": "テスト1"
-        },
-        {
-            "date": "02",
-            "start": "09:00",
-            "end": "18:00",
-            "break": "1:00",
-            "remark": "テスト2"
-        },
-        {
-            "date": "03",
-            "start": "09:00",
-            "end": "18:00",
-            "break": "1:00",
-            "remark": "テスト3"
-        }
-    ])
-
 if __name__ == '__main__':
     app.run(debug=True)
